+++ library/feature/theme/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/feature/theme/di/ThemeSettingsModule.kt
@@ -43,7 +43,7 @@ import org.koin.dsl.module
 
 val themeSettingsModule: Module = module {
     viewModel { ThemeSettingsViewModel(preferences = get()) }
-    viewModel { SeasonalThemesViewModel(seasonal = get(), theme = get()) }
+    viewModel { SeasonalThemesViewModel(seasonal = get()) }
     viewModel { SeasonalThemeOverlayViewModel(seasonal = get(), theme = get()) }
     single<SeasonalThemeManager> { SeasonalThemeManager(application = androidApplication()) }