+++ library/feature/theme/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/feature/theme/ui/seasonal/SeasonalThemeManager.kt
@@ -28,7 +28,6 @@ import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.lifecycle.setViewTreeLifecycleOwner
 import androidx.lifecycle.setViewTreeViewModelStoreOwner
 import androidx.savedstate.setViewTreeSavedStateRegistryOwner
-import com.mihaicristiancondrea.android.libs.apptoolkit.core.designsystem.ui.style.AppTheme
 import com.mihaicristiancondrea.android.libs.apptoolkit.feature.theme.R
 
 /**
@@ -128,11 +127,7 @@ class SeasonalThemeManager(
             descendantFocusability = ViewGroup.FOCUS_BLOCK_DESCENDANTS
             isFocusable = false
             isClickable = false
-            setContent {
-                AppTheme {
-                    SeasonalThemeOverlay()
-                }
-            }
+            setContent { SeasonalThemeOverlay() }
         }
 
         content.addView(